+++ examples/enum-role/src/role.rs
@@ -1,0 +1,7 @@
+
+// `PartialEq` and `Clone` is required
+#[derive(PartialEq, Clone)]
+pub enum Role {
+    ADMIN,
+    MANAGER
+}