+++ src/middleware.rs
@@ -30,6 +30,7 @@ use std::task::{Context, Poll};
 /// }
 ///
 /// // You can use both &ServiceRequest and &mut ServiceRequest
+/// // Futhermore, you can use you own type instead of `String` (e.g. Enum).
 /// async fn extract(_req: &ServiceRequest) -> Result<Vec<String>, Error> {
 ///    // Here is a place for your code to get user permissions/grants/permissions from a request
 ///    // For example from a token or database
@@ -39,23 +40,27 @@ use std::task::{Context, Poll};
 /// }
 ///
 /// // `has_permissions` is one of options to validate permissions.
+/// // `proc-macro` crate has additional features, like ABAC security and custom types. See examples and `proc-macro` crate docs.
 /// #[get("/admin")]
 /// #[has_permissions("ROLE_ADMIN")]
 /// async fn you_service() -> impl Responder {
 ///     HttpResponse::Ok().finish()
 /// }
 /// ```
-pub struct GrantsMiddleware<T, Req>
+pub struct GrantsMiddleware<E, Req, Type>
 where
-    for<'a> T: PermissionsExtractor<'a, Req>,
+    for<'a> E: PermissionsExtractor<'a, Req, Type>,
+    Type: PartialEq + Clone + 'static,
 {
-    extractor: Rc<T>,
+    extractor: Rc<E>,
     phantom_req: PhantomData<Req>,
+    phantom_type: PhantomData<Type>,
 }
 
-impl<T, Req> GrantsMiddleware<T, Req>
+impl<E, Req, Type> GrantsMiddleware<E, Req, Type>
 where
-    for<'a> T: PermissionsExtractor<'a, Req>,
+    for<'a> E: PermissionsExtractor<'a, Req, Type>,
+    Type: PartialEq + Clone + 'static,
 {
     /// Create middleware by [`PermissionsExtractor`].
     ///
@@ -72,26 +77,37 @@ where
     ///      // For example from a token or database
     ///     Ok(vec!["WRITE_ACCESS".to_string()])
     /// }
+    ///
+    /// // Or with you own type:
+    /// #[derive(PartialEq, Clone)] // required bounds
+    /// enum Permission { WRITE, READ }
+    /// async fn extract_enum(_req: &ServiceRequest) -> Result<Vec<Permission>, Error> {
+    ///     // Here is a place for your code to get user permissions/grants/permissions from a request
+    ///      // For example from a token, database or external service
+    ///     Ok(vec![Permission::WRITE])
+    /// }
     /// ```
     ///
     ///[`PermissionsExtractor`]: crate::permissions::PermissionsExtractor
-    pub fn with_extractor(extractor: T) -> GrantsMiddleware<T, Req> {
+    pub fn with_extractor(extractor: E) -> GrantsMiddleware<E, Req, Type> {
         GrantsMiddleware {
             extractor: Rc::new(extractor),
             phantom_req: PhantomData,
+            phantom_type: PhantomData,
         }
     }
 }
 
-impl<S, B, T, Req> Transform<S> for GrantsMiddleware<T, Req>
+impl<S, B, E, Req, Type> Transform<S> for GrantsMiddleware<E, Req, Type>
 where
     S: Service<Request = ServiceRequest, Response = ServiceResponse<B>, Error = Error> + 'static,
-    for<'a> T: PermissionsExtractor<'a, Req> + 'static,
+    for<'a> E: PermissionsExtractor<'a, Req, Type> + 'static,
+    Type: PartialEq + Clone + 'static,
 {
     type Request = ServiceRequest;
     type Response = ServiceResponse<B>;
     type Error = Error;
-    type Transform = GrantsService<S, T, Req>;
+    type Transform = GrantsService<S, E, Req, Type>;
     type InitError = ();
     type Future = Ready<Result<Self::Transform, Self::InitError>>;
 
@@ -100,23 +116,26 @@ where
             service: Rc::new(RefCell::new(service)),
             extractor: self.extractor.clone(),
             phantom_req: PhantomData,
+            phantom_type: PhantomData,
         }))
     }
 }
 
-pub struct GrantsService<S, T, Req>
+pub struct GrantsService<S, E, Req, Type>
 where
-    for<'a> T: PermissionsExtractor<'a, Req> + 'static,
+    for<'a> E: PermissionsExtractor<'a, Req, Type> + 'static,
 {
     service: Rc<RefCell<S>>,
-    extractor: Rc<T>,
+    extractor: Rc<E>,
     phantom_req: PhantomData<Req>,
+    phantom_type: PhantomData<Type>,
 }
 
-impl<S, B, T, Req> Service for GrantsService<S, T, Req>
+impl<S, B, E, Req, Type> Service for GrantsService<S, E, Req, Type>
 where
     S: Service<Request = ServiceRequest, Response = ServiceResponse<B>, Error = Error> + 'static,
-    for<'a> T: PermissionsExtractor<'a, Req>,
+    for<'a> E: PermissionsExtractor<'a, Req, Type>,
+    Type: PartialEq + Clone + 'static,
 {
     type Request = ServiceRequest;
     type Response = ServiceResponse<B>;
@@ -132,7 +151,7 @@ where
         let extractor = Rc::clone(&self.extractor);
 
         Box::pin(async move {
-            let permissions: Vec<String> = extractor.extract(&mut req).await?;
+            let permissions: Vec<Type> = extractor.extract(&mut req).await?;
             req.attach(permissions);
 
             let fut = service.borrow_mut().call(req);